refactor(api): move members to modules/ as controller, model, service
