+++ apps/api/src/planner.ts
@@ -2,7 +2,7 @@ import { Elysia } from 'elysia';
 import { HttpError, pgErrorCode } from './shared/lib';
 import { authContext } from './shared/auth-context';
 import { projectRoutes } from './projects/routes';
-import { memberRoutes } from './members/routes';
+import { memberRoutes } from './modules/members';
 import { roleRoutes } from './roles/routes';
 import { inviteRoutes } from './modules/invites';
 import { columnRoutes } from './modules/columns';
